requirements.txt
discord.py>=2.0.0
asyncpg>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pytz>=2023.3
pydantic>=2.0.0